restructure client
